@@ -20,15 +20,15 @@ apply plugin: 'com.github.johnrengelman.shadow'
 apply plugin: 'eclipse'
 apply plugin: 'maven-publish'
 
-version = '2.0.0-Alpha-0.1'
+version = '2.0.0-Alpha-0.2'
 group = 'com.talhanation.smallships' // http://maven.apache.org/guides/mini/guide-naming-conventions.html
-archivesBaseName = 'smallships-1.17.1'
+archivesBaseName = 'smallships-1.18.2'
 
-java.toolchain.languageVersion = JavaLanguageVersion.of(16)
+java.toolchain.languageVersion = JavaLanguageVersion.of(17)
 
 println('Java: ' + System.getProperty('java.version') + ' JVM: ' + System.getProperty('java.vm.version') + '(' + System.getProperty('java.vendor') + ') Arch: ' + System.getProperty('os.arch'))
 minecraft {
-    mappings channel: 'official', version: '1.17.1'
+    mappings channel: 'official', version: '1.18.2'
     accessTransformer = file('src/main/resources/META-INF/accesstransformer.cfg')
 
     runs {
@@ -78,11 +78,11 @@ minecraft {
 sourceSets.main.resources { srcDir 'src/generated/resources' }
 
 dependencies {
-    minecraft 'net.minecraftforge:forge:1.17.1-37.1.1'
+    minecraft 'net.minecraftforge:forge:1.18.2-40.1.54'
 
-    shadow fg.deobf("de.maxhenkel.corelib:corelib:1.17.1-1.0.2:api")
-    runtimeOnly fg.deobf("de.maxhenkel.corelib:corelib:1.17.1-1.0.2")
-    runtimeOnly fg.deobf("de.maxhenkel.corelib:corelib:1.17.1-1.0.2:javadoc")
+    shadow fg.deobf("de.maxhenkel.corelib:corelib:1.18.2-1.0.3:api")
+    runtimeOnly fg.deobf("de.maxhenkel.corelib:corelib:1.18.2-1.0.3")
+    runtimeOnly fg.deobf("de.maxhenkel.corelib:corelib:1.18.2-1.0.3:javadoc")
 
 }
 repositories {
