@@ -12,10 +12,7 @@ import net.minecraft.client.model.EntityModel;
 import net.minecraft.client.model.geom.ModelLayerLocation;
 import net.minecraft.client.model.geom.ModelPart;
 import net.minecraft.client.model.geom.PartPose;
-import net.minecraft.client.model.geom.builders.CubeListBuilder;
-import net.minecraft.client.model.geom.builders.LayerDefinition;
-import net.minecraft.client.model.geom.builders.MeshDefinition;
-import net.minecraft.client.model.geom.builders.PartDefinition;
+import net.minecraft.client.model.geom.builders.*;
 import net.minecraft.resources.ResourceLocation;
 
 public class ModelCannonBall extends EntityModel<CannonBallEntity> {
@@ -32,7 +29,19 @@ public class ModelCannonBall extends EntityModel<CannonBallEntity> {
 		MeshDefinition meshdefinition = new MeshDefinition();
 		PartDefinition partdefinition = meshdefinition.getRoot();
 
-		PartDefinition cannonball = partdefinition.addOrReplaceChild("cannonball", CubeListBuilder.create(), PartPose.offset(0.0F, 23.5F, 0.0F));
+		PartDefinition cannonball = partdefinition.addOrReplaceChild("cannonball", CubeListBuilder.create()
+				.addBox(1.5F, -4.0F, -1.0F, 1.0F, 2.0F, 2.0F)
+				.addBox(-3.0F, -1.0F, -2.0F, 4.0F, 1.0F, 4.0F)
+				.addBox(-3.0F, -6.0F, -2.0F, 4.0F, 1.0F, 4.0F)
+				.addBox(-2.0F, -6.5F, -1.0F, 2.0F, 1.0F, 2.0F)
+				.addBox(-2.0F, -0.5F, -1.0F, 2.0F, 1.0F, 2.0F)
+				.addBox(-3.0F, -5.0F, 2.0F, 4.0F, 4.0F, 1.0F)
+				.addBox(-2.0F, -4.0F, 2.5F, 2.0F, 2.0F, 1.0F)
+				.addBox(-2.0F, -4.0F, -3.5F, 2.0F, 2.0F, 1.0F)
+				.addBox(-3.0F, -5.0F, -3.0F, 4.0F, 4.0F, 1.0F)
+				.addBox(1.0F, -5.0F, -2.0F, 1.0F, 4.0F, 4.0F)
+				.addBox(-4.0F, -5.0F, -2.0F, 1.0F, 4.0F, 4.0F)
+				.addBox(-4.5F, -4.0F, -1.0F, 1.0F, 2.0F, 2.0F), PartPose.offset(0.0F, 23.5F, 0.0F));
 
 		return LayerDefinition.create(meshdefinition, 16, 16);
 	}
