@@ -7,7 +7,6 @@ import com.mojang.math.Vector3f;
 import com.talhanation.smallships.Main;
 import com.talhanation.smallships.client.model.ModelCannon;
 import com.talhanation.smallships.entities.AbstractCannonShip;
-import com.talhanation.smallships.entities.CogEntity;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.texture.OverlayTexture;
 import net.minecraft.resources.ResourceLocation;
@@ -27,9 +26,9 @@ public class RenderCannon{
         matrixStackIn.scale(0.75F, 0.75F, 0.75F);
 
         ModelCannon model = new ModelCannon();
-        model.setupAnim((CogEntity) cannonShip, partialTicks, 0.0F, -0.1F, 0.0F, 0.0F);
-        VertexConsumer ivertexbuilder = bufferIn.getBuffer(model.renderType(getCannonTexture()));
-        model.renderToBuffer(matrixStackIn, ivertexbuilder, packedLight, OverlayTexture.NO_OVERLAY, 1.0F, 1.0F, 1.0F, 1.0F);
+        model.setupAnim(cannonShip, partialTicks, 0.0F, -0.1F, 0.0F, 0.0F);
+        VertexConsumer vertexConsumer = bufferIn.getBuffer(model.renderType(getCannonTexture()));
+        model.renderToBuffer(matrixStackIn, vertexConsumer, packedLight, OverlayTexture.NO_OVERLAY, 1.0F, 1.0F, 1.0F, 1.0F);
         matrixStackIn.popPose();
     }
 
