@@ -89,8 +89,8 @@ public class RenderEntityCog extends EntityRenderer<CogEntity> {
         matrixStackIn.translate(0.0D, -1.8D,-1.0D);
         matrixStackIn.mulPose(Vector3f.YP.rotationDegrees(-90F));
         this.model.setupAnim(entityIn, partialTicks, 0.0F, -0.1F, 0.0F, 0.0F);
-        VertexConsumer ivertexbuilder = bufferIn.getBuffer(this.model.renderType(getTextureLocation(entityIn)));
-        this.model.renderToBuffer(matrixStackIn, ivertexbuilder, packedLightIn, OverlayTexture.NO_OVERLAY, 1.0F, 1.0F, 1.0F, 1.0F);
+        VertexConsumer vertexConsumer = bufferIn.getBuffer(this.model.renderType(getTextureLocation(entityIn)));
+        this.model.renderToBuffer(matrixStackIn, vertexConsumer, packedLightIn, OverlayTexture.NO_OVERLAY, 1.0F, 1.0F, 1.0F, 1.0F);
 
         //render Banner
         entityIn.renderBanner(matrixStackIn,bufferIn,packedLightIn,partialTicks);
