@@ -28,13 +28,11 @@ import net.minecraft.util.Mth;
 import net.minecraft.world.phys.Vec3;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.level.Level;
-import net.minecraftforge.fmllegacy.network.NetworkHooks;
+import net.minecraftforge.network.NetworkHooks;
 
 import javax.annotation.Nullable;
 
 public class BriggEntity extends AbstractCannonShip{
-
-    private static final EntityDataAccessor<Integer> CARGO = SynchedEntityData.defineId(BriggEntity.class, EntityDataSerializers.INT);
 
     public BriggEntity(EntityType<? extends BriggEntity> type, Level world) {
         super(type, world);
@@ -55,7 +53,6 @@ public class BriggEntity extends AbstractCannonShip{
 
     public void tick() {
         super.tick();
-        initInventory();
     }
 
     @Override
@@ -68,38 +65,11 @@ public class BriggEntity extends AbstractCannonShip{
         return 1.75D;
     }
 
-    ////////////////////////////////////DATA////////////////////////////////////
-
-    @Override
-    protected void defineSynchedData() {
-        super.defineSynchedData();
-        entityData.define(CARGO, 0);
-    }
-
-    ////////////////////////////////////SAVE DATA////////////////////////////////////
-
-    @Override
-    public void addAdditionalSaveData(CompoundTag nbt) {
-       super.addAdditionalSaveData(nbt);
-        nbt.putInt("Cargo", getCargo());
-    }
-
-    @Override
-    public void readAdditionalSaveData(CompoundTag nbt) {
-        super.readAdditionalSaveData(nbt);
-        this.setCargo(nbt.getInt("Cargo"));
-    }
+    ////////////////////////////////////GET////////////////////////////////////
 
     @Override
     public double getShipDefense() { //in %
         return 30;
-    }
-
-
-    ////////////////////////////////////GET////////////////////////////////////
-
-    public int getCargo() {
-        return entityData.get(CARGO);
     }
 
     @Override
@@ -152,11 +122,6 @@ public class BriggEntity extends AbstractCannonShip{
         return 8;
     }
 
-    ////////////////////////////////////SET////////////////////////////////////
-
-    public void setCargo(int cargo){
-        entityData.set(CARGO, cargo);
-    }
     ////////////////////////////////////INTERACTIONS///////////////////////////////
 
     @Override
@@ -231,7 +196,7 @@ public class BriggEntity extends AbstractCannonShip{
                 public AbstractContainerMenu createMenu(int i, Inventory playerInventory, Player playerEntity) {
                     return new BasicShipContainer(i, BriggEntity.this, playerInventory);
                 }
-            }, packetBuffer -> {packetBuffer.writeUUID(getUUID());});
+            }, packetBuffer -> packetBuffer.writeUUID(getUUID()));
         } else {
             Main.SIMPLE_CHANNEL.sendToServer(new MessageOpenGui(player));
         }
@@ -473,33 +438,11 @@ public class BriggEntity extends AbstractCannonShip{
                 }
             }
             f = f - 0.5;
-            Vec3 vector3d = (new Vec3((double)f, 0.0D, 0.0D + d)).yRot(-this.getYRot() * ((float)Math.PI / 180F) - ((float)Math.PI / 2F));
-            passenger.setPos(this.getX() + vector3d.x, this.getY() + (double)f1, + this.getZ() + vector3d.z);
+            Vec3 vector3d = (new Vec3(f, 0.0D, 0.0D + d)).yRot(-this.getYRot() * ((float)Math.PI / 180F) - ((float)Math.PI / 2F));
+            passenger.setPos(this.getX() + vector3d.x, this.getY() + (double)f1, this.getZ() + vector3d.z);
             passenger.setYRot(passenger.getYRot() + this.deltaRotation);
             passenger.setYHeadRot(passenger.getYHeadRot() + this.deltaRotation);
-            applyYawToEntity(passenger);
+            applyOriantationsToEntity(passenger);
         }
-
-    }
-
-    public void initInventory(){
-        SimpleContainer inventory = this.getInventory();
-        int sigma, tempload = 0;
-        for (int i = 0; i < inventory.getContainerSize(); i++) {
-            if (!inventory.getItem(i).isEmpty())
-                tempload++;
-        }
-        if (tempload > 31) {
-            sigma = 4;
-        } else if (tempload > 16) {
-            sigma = 3;
-        } else if (tempload > 8) {
-            sigma = 2;
-        } else if (tempload > 3) {
-            sigma = 1;
-        } else {
-            sigma = 0;
-        }
-        entityData.set(CARGO, sigma);
     }
 }
